renamed user dir and rendered userinfo comp on friends feed
