+++ react-app/src/components/Messages/Messages.module.css
@@ -14,9 +14,4 @@
     display: flex;
     justify-content: space-between;
     align-items: center;
-    font-size: 30px;
-    font-weight: 600;
-    border-bottom: 2px solid darkgreen;
-    margin: 0 5px 0 0;
-    text-shadow: 1px 1px 1px black, 2px 2px 1px #0bda51;
 }
