--- react-app/src/components/User/User.module.css
@@ -1,3 +1,0 @@
-.wrapper {
-    display: flex;
-}
