--- react-app/src/components/User/index.js
@@ -1,3 +1,0 @@
-import UserInfo from "./User";
-
-export default UserInfo;
