+++ react-app/src/components/UserInfo/Avatar.js
