+++ react-app/src/components/UserInfo/UserInfo.js
@@ -2,14 +2,14 @@ import React from "react";
 
 import Avatar from "./Avatar";
 
-import styles from "./User.module.css";
+import styles from "./UserInfo.module.css";
 // className={styles. }
 
 export default function UserInfo(props) {
     return (
         <div className={styles.wrapper}>
             <Avatar user={props.user} />
-            <div>{props.user.username}</div>
+            <div className={styles.username}>{props.user.username}</div>
         </div>
     );
 }
